MB-31002: Add protocol spec for auth through external process
Memcached may be configured to use an external process as an AUTH provider.
The provider needs to register itself by sending `AuthProvider` on a
connection authenticated as `@ns_server` with `Duplex` enabled. Once
successfully registered memcached will start using the connection. It is not
defined how memcached will utilize multiple registered connections (that's an
internal implementation detail within memcached). It may use one, it may
round robin etc. It is recommended that multiple connections is registered
to be prepared for failure situations.

The authentication protocol between memcached and external RBAC providers
proxies the SASL protocol between the external client and memcached.

Change-Id: I675a69d4e1bac875061ba1a7e433c3cd2d221e5a
Reviewed-on: http://review.couchbase.org/98633
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Timofey Barmin <timofey.barmin@couchbase.com>