RBAC: Add initial support for external authentication
This is a proof of concept implementation of supporting
external authentication from memcached.

There is no timeout of authentication requests. Once sent
to the external auth server we'll wait "forever" until the
reply is received, or the connection to the authentication
provider is disconnected. If the latter is true ETMPFAIL is
returned to the client (the authentication request is _not_
automatically redistributed to another authentication
provider).

The design in the prototype is pretty complex to avoid possible
deadlocks between the various threads, but it works the following
way.

A connection receives SASL_AUTH, and creates a task to be scheduled
on the executor threads and sets the connection in a suspended state
to allow other connections to use the frontend worker thread.

Some time later the executor thread picks up this task and schedule
it to run on one of the executor thrads. It starts performing the
SASL Authentication, and if it fails with "no user" AND the
mechanism is PLAIN it adds itself to the the external auth manager
and suspend itself.

Some time later the auth manager picks up the list of sasl tasks
and looks up the the active providers. If none exists it notifies
all of the tasks with a failure indicating that there isn't an
external provider available. If one exits, the worker thread
it belongs to is locked, before we incject an authentication
request for all of the pending authentication tasks.

Some time later the connection where we just injected all of these
tasks is scheduled to run, and it'll send all of these requests
to the external provider, and pick up the responses and go
the same route back.

Change-Id: I94545e118cb020bfd0496ea6e99599018a2aee43
Reviewed-on: http://review.couchbase.org/98787
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>