MB-30865: [5/n] Change DocKey API to prepare for leb encoded keys
This patch does a simple API change so that the bulk of the
leb128 (in-memory) patch is smaller.

The primary change is that DocKey (which is a view of a key)
needs to know if the key it is viewing encodes a collection-ID.

So we no longer are able to set the Collection-ID via the constructor,
we simply state if the key we are viewing does or does not
encode a collection-ID. However because this patch doesn't introduce
in-memory encoded keys, internally the DocKey constructor just sets
the Collection-ID to the default collection. However any where in
the code that does require a collection-ID other than default, the
patch adds some temporary constructors so that the ID can be specified.

The next patch in this series, removes those temporary consructors
and up-dates DocKey/StoredDocKey/Serialised/DocKey to use the
encoded flag to properly store and decode a leb128 prefixed key.

Change-Id: I11d0c5e2c6a2290a6ac5968d072c66acab99e288
Reviewed-on: http://review.couchbase.org/98758
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>