MB-30302: Simplification of Stream logging
In active/notifier/passive streams we currently have a global
logger used when a connection has been deleted before a stream.
This overcomplicates the shutdown of buckets that needs to be
done in future patch sets (registration of BucketLoggers) as
ep_engine.cc and ep_unit_tests_main would need to include
Stream files to properly destruct all global loggers.

Instead, use the globalBucketLogger and manually append the
required prefix before calls to the logging function.

Change-Id: I6be683c75910afbba2290b714f603304e09dd48e
Reviewed-on: http://review.couchbase.org/100360
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Chris Farman <christopher.farman@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>