MB-31071: catch std::overflow_error in "debug_helpers"
The code used a mix of checked_snprintf and snprintf as they
snprintf behaves slightly different on windows. checked_snprintf
throws std::overflow_error for errors (or running out of space)

As part of audit of illegal packets it only wants to dump
a fixed amount of the packet in the audit log, and depending
on where we ran out of space in formatting the packet dump
it _could_ get the exception thrown causing the audit to be
dropped and the connection disconnected.

Change-Id: Ib0b8eb3bfd8c31ad4c25584f39e32b2ff1b614a9
Reviewed-on: http://review.couchbase.org/98874
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>