MB-30302: Register/Unregister BucketLoggers
Register BucketLoggers on creation in the spdlog registry
within the memcached_logger library. Unregister them on
destruction.

Leverages changes made to spdlog/details/registry.h
wherein it now stores weak_ptr's to ensure that a BucketLogger
can be deleted when it is no longer needed.

This will allow us to change verbosity of all loggers at
runtime in a future patch. Loggers that set a custom verbosity
(views & fts) are unregistered on creation of the DcpProducer
object preventing their verbosity from being changed. This is
in keeping with previous behaviour and ensure we don't end up
in a situation where we cannot revert verbosity settings.

Change-Id: I118fed644418b538dcef54e46c123fef0be6ffc7
Reviewed-on: http://review.couchbase.org/100363
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>