Refactor: RevokeUserPermissions -> UpdateUserPermissions
The command should be used to update the users permission (and
possibly delete). There is no point of having ns_server call
RevokeUserPermissions to trigger memcached to fetch the updated
permissions with GetUserPermissions.

There is no tests for this for the moment, as the unit tests
will be written on top of the support for using external
authentication providers.

Change-Id: Icc79ea060505deb4fbbc5318a33d66c5e54177f3
Reviewed-on: http://review.couchbase.org/99008
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>