Remove unused server opcode GetUserPermissions
The server command was added in order to allow memcached ask the
external authentication provider about the RBAC settings for
a given user. It was never implemented in memcached, and we
later decided that the authentication response should contain
the RBAC profile for the user. The external authentication
provider would then call UpdateUserPermissions to update the
entry for the user.

Change-Id: I8322233eb74ae095dd731216e60949e69766ec05
Reviewed-on: http://review.couchbase.org/99013
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>