Revert "MB-30490: Remove passing of readhandle"
This reverts commit 752b0e616ba32a4fdf06c3e2d29ceeb171cfac5d.

When collection flush support is added we will need to
determine if items found in-memory are logicallyDeleted,
thus we need to re-instate this patch so that in-memory
items can be evaluated against the collections config.

This patch is nearly a clean 'git revert' 1 manual
change was required in backfill_disk.cc because the
Callback has changed how it accesses the manifest.

Change-Id: I2f9982e20f30328d0e1a118b9b9145383afe8a1e
Reviewed-on: http://review.couchbase.org/99030
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>