MB-30552: Convert type of vBucket id's from uint16_t to Vbid [1/n]
(Stage 1 of n) This commit aims to change the usage type of vBucket
id's to the new type 'Vbid' within the inner workings of each
function. This allows greater control over what operations can be
performed on a vBucket id, as well as ensuring the output formatting
is consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted.

This commit has been cherry-picked and then reverted, due to build
errors. It's previous changeID was
 I0454882561e4fb7a91f5405f59e9489a5cfe92f1

Change-Id: I4a507179dd08cb8909e71aabd3ed19a40938a650
Reviewed-on: http://review.couchbase.org/99309
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>