MB-30552: Convert type of vBucket id's from uint16_t to Vbid [3/n]
(Stage 3 of n) This commit changes to the type of Vbid inside a
collection of kvstore files. This allows greater control over what
operations can be performed on a vBucket id, as well as ensuring
the output formatting is consistent. To support the splitting of
this into multiple stages, Vbid can currently be implicitly
converted back to uint16_t to cover elements that have not yet
been converted. During this, there may be logs or prints with
"vb:vb:", which should be fixed within this set of changes.

Change-Id: I25d66ffd1e45cffdbf0af5eec713bc8924958c5d
Reviewed-on: http://review.couchbase.org/99133
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>