Refactor: Change confusing completePhase variable name
The PagingVisitor class contains a boolean variable entitled
"completePhase" which is used to indicate that sufficient documents
have been evicted and memory usage has fallen below the low watermark
so the eviction process can stop.

However when completePhase is true it indicates that the eviction
process needs to continue, and consequently the eviction process stops
only when completePhase is false.

This is confusing therefore this patch changes the name of the variable
to the more accurate name isBelowLowWaterMark and inverts the logic so
false mean we must continue with the eviction process.

Change-Id: I3a2f57752155e9a47f5a87ab6c202be2c1ba7e97
Reviewed-on: http://review.couchbase.org/99138
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>