MB-30552: Convert type of vBucket id's from uint16_t to Vbid [5/n]
(Stage 5 of n) This commit changes to the type of Vbid based
around the engine.h file. Some of these files use engine.h as a
base to override, hence the slightly wide spread in one patch.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: I9746fa86cbd6186a35376884846e566b810307ee
Reviewed-on: http://review.couchbase.org/99148
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>