MB-30913: Rename members/methods in preparation for scopes
Many collections class members/methods are named such that
it only makes sense for them to deal with 1 type of object.
To simplify future patches, rename these members/methods so that
class interfaces deal with collections/scopes more consistently.

Leave the begin and end iterator methods of Manifest as we
generally wish to iterate over all collections of a Manifest.

Change-Id: Id5b4128056d48800adb956f0c8bb780a8c3e19de
Reviewed-on: http://review.couchbase.org/99235
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>