MB-30552: Convert type of vBucket id's from uint16_t to Vbid [7/n]
(Stage 7 of n) This commit changes to the type of Vbid for
files in the second alphabetical half of src/.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: Ided697ce01ccefe5615802cff46c2070e18fee1e
Reviewed-on: http://review.couchbase.org/99167
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>