MB-31149: Fix regression where returned cas was set to 0
When sending an Append (0x0e) request to the server I'm seeing a status
of success with a CAS value of 0 when the MutationSeqNo Hello Feature is
set. When the MutationSeqNo Hello Feature is not set then the CAS value is
correct and everything looks fine.

Change-Id: If9610716a31074bcb15d65675baf9c8e60988ad0
Reviewed-on: http://review.couchbase.org/99218
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>