MB-31162: Ensure DCP Consumer sets the frequency counter
When using the new hifi_lru hash table eviction policy the frequency
counter is correctly sent by the DCP producer to its respective
consumer.  However unfortunately the consumer does not pick-up the
value, and instead uses the default value (initialFreqCount) of 4.

This patch fixes the issue by correctly setting the frequency value
of the consumer item to the value that it receives from the producer.

Change-Id: Id93df0bd26f85e251037181abaeec6e8497dedd1
Reviewed-on: http://review.couchbase.org/99416
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>