MB-30552: Convert type of vBucket id's from uint16_t to Vbid [10/n]
(Stage 10 of n) This commit changes to the type of Vbid for
files in daemon.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: I09bce73b06987f211543ee92facf0d4139069799
Reviewed-on: http://review.couchbase.org/99629
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>