MB-30552: Convert type of vBucket id's from uint16_t to Vbid [12/n]
(Stage 12 of n) This commit changes to the type of Vbid for
files inside protocol.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: Id6ee3e367ffc9f006f9ed75c4d9c34e638525e58
Reviewed-on: http://review.couchbase.org/99631
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>