MB-30552: Convert type of vBucket id's from uint16_t to Vbid [13/n]
(Stage 13 of n) This commit changes to the type of Vbid for
test files, including testapp and testsuite.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: I487e96432cbf885693c8bb459ba1e269f191691f
Reviewed-on: http://review.couchbase.org/99632
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>