MB-26618: [3/n] Remove optional JSON parameter from open
DCP open now does not accept the optional value. This
patch also incorporates MB-24547, the code which would
force replication streams into being collection-aware.
ns_server now performs that functionality and since
the JSON value was part of that temporary code, the
dcpOpen retry logic is dropped in this patch.

Change-Id: Iae1cf8cf0f390e44b6a25dac420b2fd2d4247bb5
Reviewed-on: http://review.couchbase.org/99374
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>