MB-31141: Don't reject snappy|raw DCP deletes
A related bug means that is possible for 5.x to create
deletes with a non-zero raw value. When 5.5x backfills such
an item for transmission to another 5.5x node (and snappy
is enabled), the delete gets sent with a snappy datatype
and rejected by the target node, which triggers a rebalance
failure.

Change-Id: Ib91453f96882ef4e01ee0e2a3e5e51ed0786a325
Reviewed-on: http://review.couchbase.org/99414
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>