MB-31208:  Don't reset frequency counter on a mutation
Each document in the hash table has a frequency counter asscoiated with
it.  The counter is incremented each time the document is read.

Currently if a document is assigned a new value the frequency counter
is reset back to the initialFreqCount of 4.  This is a bug; the
frequency counter should remain unchanged, and is fixed in this patch.

Note:  The original 2-bit_lru eviction policy has the correct behaviour
in that its nru value is not reset when the document is assigned a new
value.

Change-Id: I90060e7b4b207d65502b846d20f71e4b252f8800
Reviewed-on: http://review.couchbase.org/99479
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>