MB-30552: Add test for correct log formatting of Vbid
Added a test to ensure that Vbid logs correctly, along with a fix
to make that actually happen. <spdlog/fmt/ostr.h> is needed to
output custom types, however having this in vbucket.cc allows the
implicit conversion of Vbid to uint16_t to take precedence and
ignore the formatting. By including it in the appropriate
locations, this issue can be avoided. This also removes the need
to have a separate vbucket.cc file.

Change-Id: I7cff526218c8700f5e86cf95676e4ab67fc8bdbf
Reviewed-on: http://review.couchbase.org/99439
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>