MB-30552: Convert type of vBucket id's from uint16_t to Vbid [11/n]
(Stage 11 of n) This commit changes to the type of Vbid for
remaining instances in include.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: Ie28af2fc5cc1e9d806c1c8c0eb5a7e35544a92f3
Reviewed-on: http://review.couchbase.org/99630
Reviewed-by: David Haikney <david.haikney@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>