MB-30552: Convert type of vBucket id's from uint16_t to Vbid [14/n]
(Stage 14 of n) This commit changes to the type of Vbid for
ep benchmark files.
Vbid allows greater control over what operations can be performed
on a vBucket id, as well as ensuring the output formatting is
consistent. To support the splitting of this into multiple stages,
Vbid can currently be implicitly converted back to uint16_t to
cover elements that have not yet been converted. During this,
there may be logs or prints with "vb:vb:", which should be fixed
within this set of changes.

Change-Id: I4c0057dbd4edd1a119d45e9ca59a6c726c479c72
Reviewed-on: http://review.couchbase.org/99633
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>