MB-31326: Accept UID in collections stream-request
Update DCP stream-request so that a collection enabled client can
encode a collection manifest UID.

Why?

A collection manifest UID is required for resuming collection-enabled
DCP streams. The manifest UID the client gives as part of stream
request is used to verify that the client can resume safely following
disconnect (in particular fail-over). A specific scenario is when
client notes a collection was dropped and then failure forces the
client to use a newly promoted vbucket, who has yet to see the
collection drop. In this case, the client would be held off from
connecting to the new active until the new active has picked up the
new manifest from the cluster manager.

With collections a DCP client can be either streaming everything (i.e.
replication) or filtering (index building on a single colllection).
The following describes the flow of stream-requests with respect to
the manifest-uid.

Client wants everything:

A client begins by stream requesting a VB from seqno 0. In this case
the client has no knowledge of the VB and thus no manifest-UID, the
client stream-requests with no JSON value.

Next following a disconnect, the client will want to resume
the stream, so a non-zero start seqno is specified and is required to
specify the last manifest-UID they received.

  {"uid":"71"}

As per other collection JSON specifications, the UID is a base-16
value with no 0x prefix (precious bytes!).

Note: the client could still specify no JSON value if the stream
hasn't sent any collection manifest UID yet. KV does not error check
this scenario and it is the clients responsibility to resume a stream
correctly.

Client is filtering:

A client begins by stream requesting a VB from seqno 0. In this case
the client has no knowledge of the VB and thus no manifest-UID,
however the client is configuring filtering so encodes a JSON document
with the filter configuration as the stream-request value.

  {"collections":["a1", "7", "c21"]}

Next following a disconnect, the client will want to resume
the stream, so a non-zero start seqno is specified and is required to
specify the last manifest-UID they received.

  {"uid":"71", "collections":["a1", "7", "c21"]}

Change-Id: I824b953b9273cae6a32eb3c8395f2491ea44b018
Reviewed-on: http://review.couchbase.org/99676
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>