MB-31327: Ephemeral backfill does not respect purge seqno.
There exists a bug in Ephemeral backfill_memory such that we can
start a backfill for a startSeqno > purgeSeqno of a given vBucket.
This leads to data inconsistencies as we may not replciate a delete.

Fix the bug by checking that the purgeSeqno < startSeqno of a
backfill when startSeqno != 1.

Only fix for buffered memory backfill as the non-buffered backfill
is not used and will be deleted in master.

Change-Id: Ia91d6d1e82fcfb7ea2d6463c76a531ea0d8aa182
Reviewed-on: http://review.couchbase.org/99713
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: David Haikney <david.haikney@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>