MB-31175: Ephemeral HTTombstonePurger underflows age check
There exists a bug in the HTTombstonePurger such that we can cause
an integer underflow when determining if a deleted item should be
marked stale (for removal). This occurs because the task records and
uses a start time when it is created. When the task runs slowly due
to heavy write locking on the seqList an item can be deleted seemingly
in the future to the HTTombstonePurger. This causes an integer
underflow when we perform the "now - osv->getDeletedTime() >= purgeAge"
check and the subsequent removal of the tombstone before the purgeAge.

Fix this bug by checking that now >= the deleted time of an osv.

Change-Id: Ibffdb88820d5335dfb7b75c2cc250fc766847cf7
Reviewed-on: http://review.couchbase.org/99744
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: David Haikney <david.haikney@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>