MB-31127: Use a dedicated RBAC database for external users
Previously we used a common RBAC database for users defined in
the local domain and users defined in the external domain. All
entries was located in the RBAC file and transferred to memcached
by having memcached reloading the database file.

As part of adding support for external authentication though
ns_server new RBAC entries is part of the response from ns_server.
This introduced the problem that these entries would be lost if
someone tried to reload the database.

One of the current plans is that ns_server will push a new external
database in a regular interval. To avoid having to try to merge
the databases every time one of them changes this change introduce
a dedicated RBAC database for the local users, and one RBAC database
for all external users.

Change-Id: Ic6e88005af3b930922e5ae6779edb59f1aad0e8c
Reviewed-on: http://review.couchbase.org/99791
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>