Refactor: Use cb::mcbp::Status in DCP interface
The status code passed in the DCP interface was put
directly into the status field in the response message
and passed along as uint8_t. We have an enum class
representing all of the valid response codes and getter/
setter methods in the response class to perform the
correct byteswapping.

Change-Id: I04514296bcee268b25bb539be0964aa87ca03966
Reviewed-on: http://review.couchbase.org/99878
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>