Remove all uses of cb::strnstr()
This function was added for platforms which don't implement 'strnstr'
(i.e. Windows); however we only use it in two places; and in those two
places there is a better choice of a type-safe find() method on the
const_char_bufffer. As such remove the two uses (which lets us remove
the now-unused functions from platform).

Change-Id: I67bc4cddf615d4088876aee9df120565a77186d4
Reviewed-on: http://review.couchbase.org/100044
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>