Clear the packet after we've used it
Cookie::toJSON tries to add the packet currently in use
to the JSON dump which is included in the dump of the
connection. This is however not "safe" as we've consumed
the bytes from the input buffer, and we could in theory
resize the buffer and use a dangling pointer.

This isn't currently a problem (as we won't resize the
buffer until we've read the header, and then we reset
the cookie), but when using bufferevents we don't have
that control over when the underlying buffer gets reset
once we've told it that we're done with memory.

Change-Id: Ib2b56be35116c4b9b1a8addd764bc847e9754ff4
Reviewed-on: http://review.couchbase.org/100076
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>