MB-31321: Add a getFeatures function to the engine interface
We need to know if a bucket supports collections to block certain
scenarios described in the MB. To do this in an expandable way,
add a function that allows an engine to return a set of features
that it supports. This result will be stored in the memcached
Bucket object and queried in future patches.

Change-Id: I09925e35ec67bf0455fe392318b7e39b7b024c7e
Reviewed-on: http://review.couchbase.org/102803
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>