[BP] MB-31548: Fix bug in getRandomKey
This is a backport of MB-31495.

This patch fixes a bug where getRandomKey would loop indefinitely if
start == 0.

The bug is that we check for curr == start before checking if
curr == max.  See the original code below.

start = 0;
curr = start;
VBucketPtr vb = getVBucket(Vbid(curr++));
while (!vb || vb->getState() != vbucket_state_active) {
    if (curr == start) {
        return GetValue(NULL, ENGINE_KEY_ENOENT);
    }
    if (curr == max) {
        curr = 0;
    }

    vb = getVBucket(Vbid(curr++));
}

Assume the bucket is empty. The first time we enter the while loop
curr == 1.  curr != start and curr != max and so we will call
getVBucket and increment curr to 2.  We repeat until curr == 1024.
On going round the while loop again curr != start, but curr == max
and so set curr = 0.

We then call getVbucket and increment curr to 1.  On going round the
while loop again curr != start (as it is 1), and hence we loop
indefinitely.

Change-Id: I406474eb3be8565780625039c53e7d389801c1d7
Reviewed-on: http://review.couchbase.org/100320
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>