MB-31542: Fix spurious failures in dcp consumer hotness data test
The test is designed to check that eviction hotness data is picked up by
the DCP consumer.

The test sends a hotness value of 128 to the consumer.  Therefore we
would expect the consumer to have the value of 128 - which is what the
test checks.

However sometimes the value is 129.  The reason for this is that
we have to perform a get to extract the item's hotness value (so we can
check it), and as the get is an access of the item it tries to increment
the hotness value.

As we use a probabilistic counter the get sometimes causes the value to
increase to 129, whilst othertimes does not.  Therefore in the test it
is acceptable for the hotness value of the item to be either 128 or 129.

Change-Id: I937b7172ca1a0cd6e5c28f78ad30712c04cae81f
Reviewed-on: http://review.couchbase.org/100367
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>