MB-31410: Fix buffer emptying in PassiveStream::processBufferedMessages
On a DcpConsumer, when we hit the replication threshold we buffer the
following incoming messages for deferred processing in the
DcpConsumerTask.
The Consumer cannot process new incoming message until the buffer has
been completely drained. That is for ensuring that messages are
processed in the correct order, e.g.:
- Mutations are processed ordered by seqno
- SnapshotMarkers are processed before the actual snapshot Mutations
- etc..

In MB-20649 we have modified the way we synchronize the Consumer
front-end thread and the bg-thread that runs the DcpConsumerTask. That
makes possible that the front-end thread processes new incoming
mutations before the DcpConsumerTask has drained the buffer.

To fix, we remove the last message from the buffer only after we have
processed the message in the DcpConsumerTask.
That ensures that the front-end thread processes new incoming messages
only when all the buffered ones have been processed.

Change-Id: I610d8510e68867ad0abd4ce4aefd512574891e25
Reviewed-on: http://review.couchbase.org/100497
Reviewed-by: Daniel Owen <owend@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>