MB-31581: Fix missing .get() for formatting of backfillItems.sent stat
As part of the Vbid refactoring (MB-30552), many of the stats that
use vBucket ID's had .get() added to ensure that the formatting
remained consistent with before the type of Vbid was changed. One of
these stats, backfillItems.sent, was missed and thus would be
incorrect, hence the fix of adding the .get() now.

Change-Id: Ica9cede8a7ea7288858f992b7377572eb0f16a89
Reviewed-on: http://review.couchbase.org/100410
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Daniel Owen <owend@couchbase.com>