MB-31542: Ensure if decayer task runs it has no effect
It is possible for the ItemFreqDecayerTask to be scheduled in-between
the call to dcp->mutation and the get.  This would result in the
frequency value being reduced from the original value of 128.

To avoid this being a problem we set the item_freq_decayer_percent
parameter to 100, which means that if the itemFreqDecayerTask runs it
will have no effect on an item's frequency count.

Change-Id: I63f7676925b4e50261d39df6af218db086014504
Reviewed-on: http://review.couchbase.org/100450
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>