MB-30041: Add jsonGet function to json utilities
As part of MB-31599, we want to cleanup some of the error
messages output in audit. As we also wish to migrate to
nlohmann json and, where possible, use the raw nlohmann json
syntax. For simplicity, we are printing nlohmann json exceptions
for any json parsing errors. This is okay, but type errors do not
print the key of the object which we are testing.

Add a function that wraps around "nlohmann::json.get<T>()" that
catches the exceptions thrown and prepends the key to the exception
message so that end users can more easily identify issues in their
json configurations.

To keep syntax as simple as possible, this function should also
wrap the "nlohmann::json.at(key)" function, as our most common use
case is:
- "nlohmann::json.at(key).get<T>()"
These cases can then be replaced with:
- "cb::jsonGet<T>(nlohmann::json, key)"

Type errors using the pure nlohmann json method are as follows:
- "[json.exception.type_error.302] type must be <expectedType>,
 but is <actualType>"
Type errors using the cb::getJson function are as follows:
- "[json.exception.type_error.302] value for key "<key>" -
 [json.exception.type_error.302] type must be <expectedType>,
 but is <actualType>"

Change-Id: I70d609e9546173dcf9bf28c46e445f04f6504147
Reviewed-on: http://review.couchbase.org/100496
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>