MB-31619: Print correct timezone in log messages
As part of the migration to spdlog as our logger, the timezone offset
was incorrectly omitted from the log timestamp - e.g.

    2018-10-10T15:23:49.565757Z INFO Pushing new cluster config for bucket:[testB] revision:[4179]

Note the actual timestamp (15:23:49) /is/ in the local timezone (not
UTC); so this is potentially very misleading it's *not* a Zulu (UTC+0)
timestamp.

Compare this to ns_server which correctly prints the offset (-10:00 here):

    [ns_server:debug,2018-10-11T01:16:47.682-10:00,ns_1@127.0.0.1:compaction_new_daemon<0.419.0>:compaction_new_daemon:process_compactors_exit:1352]Finished compaction iteration.

Fix by inserting the missing timezone offset formatter.

Change-Id: I4a55396a2b73532b4cf0a2e05d0a5d0af942f45c
Reviewed-on: http://review.couchbase.org/100506
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>