MB-26907: Update expiration protocol to include delete time
This change is due to the requirement that if DCP control has been
set to output expiry opcodes, it should also have delete times
enabled which is the trigger for using deletion_v2. A number of
helper functions have been added to reduce code duplication. This
patch also moves DeleteSource into memcached for use inside daemon.

Change-Id: I9a5e2a8747249b91408258b1103d3a77ac6238d8
Reviewed-on: http://review.couchbase.org/100704
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>