Push active user list
Push the active users list to the external authentication
providers at a configurable interval. This allows the
external authentication provider to check with LDAP to
see if the users credential has changed.

The interval is specified in seconds and may be changed
dynamically by using the configuration paramater named:
active_external_users_push_interval

This commit removes the user command to fetch the list
of active users as it is no longer needed.

Change-Id: I99bd3bfa3444c29dd3170a538df6d3c148dbd707
Reviewed-on: http://review.couchbase.org/100319
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>