MB-31664: Add ability to disable use of disk backfill queue
Before removing the disk backfill queue, we want to disable its use.
This allows use to extensively test and if an issue is found we can
easily revert to the previous approach of using the backfill queue.

This patch provides the ability to disable the use of the queue with a
dynamic configuration setting.  (It is dynamic so can be switched during
testing but the end user cannot change.)  A future patch will set the
default configuration to false so the disk backfill queue is not used.

Once we are happy that there are no adverse effects with not using the
disk backfill queue it, together with the configuration switch, will be
removed.

Change-Id: Ie8add095cd7eb68b28c2728b30ed37012e40c19f
Reviewed-on: http://review.couchbase.org/100716
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>