Refactor: Create a consumer MutationResponse
The DcpProducer will never create 'extended-meta-data'
so every MutationResponse is storing a nullptr (8-bytes).

Further testing would be required to be confident that the
consumer never receives extended-meta-data, so for this
patch it still stores and handles it.

Overall this patch (and the prior patch adjusting the
sizes of our Yes/No types) reduces the size of MutationResponse
from 48 bytes to 32 bytes, and 1 MutationResponse is created
for every mutation DCP transfers (and that is per producer).

Change-Id: Ib74ef3cf58803bbd56afb4b349d648ed4bde5a56
Reviewed-on: http://review.couchbase.org/100754
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>