Add an option to bypass all of the ep engine tests
The test suites in ep-engine is slow and time consuming (given that
we run them with a lot of different configurations). By default all
tests should be built and run, but people who don't change that piece
of the code should have a way of skipping all of those tests duing
the development test cycle.

In order to bypass running all of these test suites the following
parameter should be passed to cmake:

    -DEP_BYPASS_TESTSUITES=ON

Doing so reduce the number of tests from 180 to 130, and the total
test time from 866 sec to 190 sec (running make test on my mac
laptop)

The memcached end-to-end tests add an additional slowdown due
to the 1 sec shutdown of the memcached process (and the fact
that we have so many test suites). I've created MB-31676
to address that (https://issues.couchbase.com/browse/MB-31676)

Change-Id: Id88d3224761fbb269f2d4d7ab2c5f8c5380a0ab3
Reviewed-on: http://review.couchbase.org/100757
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>