MB-31978: Add DcpStreamId into DCP API
DCP producer API relates to those methods which will send a message from
server to client only for DCP producer connections.

* snapshot-marker
* mutation
* deletion
* expiration
* system event
* stream-end

This commit adds to those function prototypes the DcpStreamId parameter
so that the DcpProducer can pass down the ID of the stream that
generated the message and will later on be included in the outgoing
packet.

From client to server the close_stream method can accept a DcpStreamId
so the client can close a stream by vb/sid identification.

Note that flow-control (ack) is not stream level, clients ack to the
producer so we never consider stream-id in any acking functions.

Change-Id: I81a45bf01340821e9e38079dbc9c2520145e6232
Reviewed-on: http://review.couchbase.org/102197
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>