MB-31688: Remove the use of [cb::]ProcessClock
It is defined as:

    using ProcessClock = std::chrono::steady_clock

The name is shorter, but it adds an extra lookup to figure out
what it is. It used to be different things when not all of the
compilers supported std::chrono::steady_clock, but now that it
is no longer the case it should be removed.

Change-Id: Iebef35e07ec2e0d62495853434f16156b7cde529
Reviewed-on: http://review.couchbase.org/100802
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>