MB-26907: Introduce usage of Expiration event for DCP
This commit introduces implementation of Expiration event within
DCP, which in turn triggers the Expiration code path instead of
the deletion path. This is done by passing the DeleteSource to
'toItem' from any code that expires a document.
This also implements the ability to output an expiry opcode from
both memory and backfill, as demonstrated by the tests included
in this patch.

Change-Id: I2809941461310fce14e5fae5786a5d2b17b87f61
Reviewed-on: http://review.couchbase.org/100561
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>